pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monarch-forecast"
version = "1.0.0"
description = "Financial forecasting desktop app powered by Monarch Money"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "flet>=0.84.0",
    "monarchmoneycommunity>=1.3.0",
    "keyring>=25.0.0",
    "flet-charts>=0.84.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.29",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25.0",
    "pre-commit>=4.0",
    "vulture>=2.14",
    "tach>=0.34",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
monarch-forecast = "src.main:run"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "RUF001", # ambiguous unicode chars — intentional for UI display (e.g. − minus sign)
    "SIM105", # contextlib.suppress — style preference, try/except is fine
    "TCH",    # TYPE_CHECKING imports — can break runtime with Flet
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ty.rules]
# Flet ships py.typed as of 0.80 and strongly-typed event generics as of
# 0.84, so we treat ty's findings as real. All Flet-facing handlers use
# `ft.Event[SpecificControl]` rather than the legacy `ft.ControlEvent`.
# If a new rule starts firing on Flet stub gaps, open a ticket on
# flet-dev/flet rather than demoting the rule here.
unresolved-import = "warn"

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
# The vulture whitelist is synthetic — it references names from other modules
# to suppress vulture false-positives, so it looks like unresolved-reference
# noise to every other tool.
exclude = ["vulture_whitelist.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["src/__init__.py", "src/**/__init__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.vulture]
# 80% confidence skips the noisy Flet/mock attribute false positives at 60%
# (things like `on_click`, `return_value`, `content` — consumed by the Flet
# runtime or mock framework, not by Python code) without needing a sprawling
# whitelist. Genuine dead code still surfaces at 80%+.
paths = ["src", "tests", "vulture_whitelist.py"]
min_confidence = 80
